+++ Core/PeekabooCore/Sources/PeekabooBridge/PeekabooBridgePayloads.swift
@@ -184,7 +184,7 @@ public struct PeekabooBridgeTargetedClickRequest: Codable, Sendable {
 
     public static func requiresPostEventPermission(target: ClickTarget, clickType: ClickType) -> Bool {
         switch (target, clickType) {
-        case (.coordinates, _), (_, .double):
+        case (.coordinates, _), (_, .double), (_, .longPress):
             true
         case (_, .single), (_, .right):
             false